requirements.txt
chroma-hnswlib==0.7.3
chromadb==0.4.22
config==0.5.1
GitPython==3.1.40
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
keras==2.15.0
kubernetes==29.0.0
langchain==0.1.0
langchain-community==0.0.11
langchain-core==0.1.8
langdetect==1.0.9
langsmith==0.0.79
numpy==1.26.3
openai==1.7.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pip==23.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.1.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
rapidfuzz==3.6.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.0
unstructured==0.11.8
unstructured-client==0.15.2
websocket-client==1.7.0
streamlit-chat==0.1.1
PyMuPDF==1.23.16
faiss-cpu==1.7.4
st-pages==0.4.5
matplotlib==3.8.2
opencv-python==4.9.0.80
openpyxl==3.1.2

"""""

pip install --upgrade --force-reinstall pymupdf

for opencv on EC2 :


    sudo yum update
    sudo yum install libglvnd-glx


    OR 

    sudo apt-get install libgl1 



pip install torchvision



"""""
